logs, editing, and saving volumes

+++ src/views/Volume.vue
@@ -69,6 +69,14 @@
         <i class="bi-trash fs-6"></i>
         Delete
       </button>
+      <button
+        class="btn btn-success ms-auto"
+        :disabled="!volumeModified"
+        @click="saveVolume()"
+      >
+        <i class="bi-save fs-6"></i>
+        Save Volume
+      </button>
     </div>
 
     <div
@@ -78,7 +86,12 @@
     >
       <table
         class="
-          table table-striped table-hover table-bordered-curved table-borderless
+          table
+          table-striped
+          table-hover
+          table-bordered-curved
+          table-borderless
+          table-sm
         "
         v-if="loaded"
       >
@@ -109,59 +122,87 @@
         </tbody>
       </table>
     </div>
+
+    <file-editor
+      ref="modal"
+      :file="openFile"
+      v-if="editing"
+      @save="saveOpenFile($event)"
+      @cancel="editing = false"
+    ></file-editor>
   </div>
 </template>
 
 <script lang="ts">
-import { VolumeI } from '@/interfaces/instances';
-import RefreshAppInfo from '@/mixins/refreshAppInfo.vue';
-import {FileSystemFileEntry, FileSystemDirectoryEntry, FileSystemDirectoryReader} from '@/interfaces/filesystem';
-import {Component, Mixins} from 'vue-property-decorator';
-import { vol as Vol, Volume as FSVolume } from 'memfs';
-import authFetch from '@/utilities/authFetch';
-import Dirent from 'memfs/lib/Dirent';
-import { TDataOut } from 'memfs/lib/encoding';
+import { VolumeI } from "@/interfaces/instances";
+import RefreshAppInfo from "@/mixins/refreshAppInfo.vue";
+import {
+  FileSystemFileEntry,
+  FileSystemDirectoryEntry,
+  FileSystemDirectoryReader,
+} from "@/interfaces/filesystem";
+import { Component, Mixins, Vue } from "vue-property-decorator";
+import { vol as Vol, Volume as FSVolume } from "memfs";
+import authFetch from "@/utilities/authFetch";
+import Dirent from "memfs/lib/Dirent";
+import { TDataOut } from "memfs/lib/encoding";
+import FileEditor from "@/components/FileEditor.vue";
+import { Modal } from "bootstrap";
 
-@Component
+@Component({
+  components: {
+    FileEditor,
+  },
+})
 export default class Volume extends Mixins(RefreshAppInfo) {
   loaded = false;
   volume!: VolumeI;
 
   fs!: typeof Vol;
 
-  currentDirectory = '/';
+  currentDirectory = "/";
   currentDirectoryListing: Dirent[] = [];
 
   selected: Dirent[] = [];
+
+  openFile: { name: string; contents: string } = { name: "", contents: "" };
+  editing = false;
+
+  volumeModified = false;
 
   listDirectory(): TDataOut[] | Dirent[] {
     return this.fs.readdirSync(this.currentDirectory, { withFileTypes: true });
   }
 
-  get currentDirectoryBreadcrumbs(): { name: string, path: string}[] {
-    const paths = this.currentDirectory.split('/');
+  get currentDirectoryBreadcrumbs(): { name: string; path: string }[] {
+    const paths = this.currentDirectory.split("/");
 
-    paths[0] = '/';
-    
-    let pathsFull: { name: string, path: string}[] = [];
+    paths[0] = "/";
+
+    let pathsFull: { name: string; path: string }[] = [];
 
     for (let i = 0; i < paths.length; i++) {
       pathsFull.push({
         name: paths[i],
-        path: paths.slice(0, i + 1).join('/').slice(1)
+        path: paths
+          .slice(0, i + 1)
+          .join("/")
+          .slice(1),
       });
     }
 
-    pathsFull[0].name = 'root';
+    pathsFull[0].name = "root";
 
     return pathsFull;
   }
 
   async getFullVolume(volumeID: string): Promise<VolumeI | null> {
-    const res = await authFetch(`${this.$store.state.adminAPIURL}/admin/getVolume?id=${volumeID}`);
-    
+    const res = await authFetch(
+      `${this.$store.state.adminAPIURL}/admin/getVolume?id=${volumeID}`
+    );
+
     if (!res.ok) return null;
-    
+
     const json = await res.json();
     return json;
   }
@@ -173,26 +214,49 @@ export default class Volume extends Mixins(RefreshAppInfo) {
       this.selected = [file];
     }
   }
-  
-  doubleClickHandler(event: MouseEvent, file: Dirent): void {
-    this.selected = [];
 
+  doubleClickHandler(event: MouseEvent, file: Dirent): void {
     if (this.isFolder(file)) {
-      if (this.currentDirectory == '/') {
-        this.currentDirectory = '/'+file.name;
+      if (this.currentDirectory == "/") {
+        this.currentDirectory = "/" + file.name;
       } else {
         this.currentDirectory = `${this.currentDirectory}/${file.name}`;
       }
+
+      this.currentDirectoryListing = this.listDirectory() as Dirent[];
+    } else {
+      const fileContents = this.fs.readFileSync(
+        `${this.currentDirectory}/${file.name}`
+      );
+
+      this.openFile = {
+        name: file.name.toString(),
+        contents: fileContents.toString(),
+      };
+
+      this.editing = true;
+
+      this.$nextTick(() => {
+        this.toggleModal(true);
+      });
     }
 
-    this.currentDirectoryListing = this.listDirectory() as Dirent[];
+    this.selected = [];
   }
 
   deleteFiles(): void {
-    if (!confirm('Are you sure you want to delete these files?')) return;
+    if (!confirm("Are you sure you want to delete these files?")) return;
 
     this.selected.forEach((file) => {
-      this.fs.unlinkSync(this.currentDirectory+'/'+file.name);
+      if (this.isFolder(file)) {
+        this.fs.rmdirSync(`${this.currentDirectory}/${file.name}`, {
+          recursive: true,
+        });
+      } else {
+        this.fs.unlinkSync(`${this.currentDirectory}/${file.name}`);
+      }
+
+      this.volumeModified = true;
     });
 
     this.selected = [];
@@ -201,28 +265,43 @@ export default class Volume extends Mixins(RefreshAppInfo) {
   }
 
   createFolder(): void {
-    const folderName = prompt('Folder Name');
+    const folderName = prompt("Folder Name");
 
-    this.fs.mkdirSync(this.currentDirectory+'/'+folderName);
+    if (!folderName) return;
+
+    this.fs.mkdirSync(this.currentDirectory + "/" + folderName);
     this.currentDirectoryListing = this.listDirectory() as Dirent[];
+
+    this.volumeModified = true;
   }
 
   createFile(): void {
-    const fileName = prompt('File Name');
+    const fileName = prompt("File Name");
 
-    this.fs.writeFileSync(this.currentDirectory+'/'+fileName, '');
+    if (!fileName) return;
+
+    this.fs.writeFileSync(this.currentDirectory + "/" + fileName, "");
     this.currentDirectoryListing = this.listDirectory() as Dirent[];
+
+    this.volumeModified = true;
   }
 
   renameFile(file: Dirent): void {
-    const newName = prompt('New name', file.name.toString());
+    const newName = prompt("New name", file.name.toString());
     if (newName == null) return;
-    this.fs.renameSync(this.currentDirectory+'/'+file.name, this.currentDirectory+'/'+newName);
+    this.fs.renameSync(
+      this.currentDirectory + "/" + file.name,
+      this.currentDirectory + "/" + newName
+    );
     this.currentDirectoryListing = this.listDirectory() as Dirent[];
+
+    this.volumeModified = true;
   }
 
   isFolder(file: Dirent): boolean {
-    return this.fs.lstatSync(this.currentDirectory+'/'+file.name).isDirectory() 
+    return this.fs
+      .lstatSync(this.currentDirectory + "/" + file.name)
+      .isDirectory();
   }
 
   async dropHandler(event: DragEvent): Promise<void> {
@@ -231,17 +310,22 @@ export default class Volume extends Mixins(RefreshAppInfo) {
     if (!event.dataTransfer) return;
 
     for (const file of event.dataTransfer.items) {
-      await this.writeFileOrDirectory(file.webkitGetAsEntry(), this.currentDirectory);
+      await this.writeFileOrDirectory(
+        file.webkitGetAsEntry(),
+        this.currentDirectory
+      );
+      this.volumeModified = true;
     }
 
     this.currentDirectoryListing = this.listDirectory() as Dirent[];
   }
 
-  async writeFileOrDirectory(file: FileSystemFileEntry | FileSystemDirectoryEntry, dir: string): Promise<void> {
-    console.log(file);
-
+  async writeFileOrDirectory(
+    file: FileSystemFileEntry | FileSystemDirectoryEntry,
+    dir: string
+  ): Promise<void> {
     if (file.isDirectory) {
-      this.fs.mkdirSync(dir+'/'+file.name);
+      this.fs.mkdirSync(dir + "/" + file.name);
 
       const directoryReader = (file as FileSystemDirectoryEntry).createReader();
       const entries = await this.readEntries(directoryReader);
@@ -249,19 +333,38 @@ export default class Volume extends Mixins(RefreshAppInfo) {
       if (entries == null) return;
 
       for (const f of entries) {
-        await this.writeFileOrDirectory(f, dir+'/'+file.name);
+        await this.writeFileOrDirectory(f, dir + "/" + file.name);
       }
     } else {
-      this.fs.writeFileSync(dir+'/'+file.name, (await file.name).toString());
+      this.fs.writeFileSync(
+        dir + "/" + file.name,
+        (
+          await (
+            await this.getFileUploadContents(file as FileSystemFileEntry)
+          ).text()
+        ).toString()
+      );
     }
+
+    this.volumeModified = true;
   }
 
-  async readEntries(directoryReader: FileSystemDirectoryReader): Promise<Array<FileSystemFileEntry | FileSystemDirectoryEntry> | null> {
+  async getFileUploadContents(file: FileSystemFileEntry): Promise<File> {
+    return new Promise((resolve, reject) => {
+      file.file(resolve, reject);
+    });
+  }
+
+  async readEntries(
+    directoryReader: FileSystemDirectoryReader
+  ): Promise<Array<FileSystemFileEntry | FileSystemDirectoryEntry> | null> {
     try {
       const entries = await new Promise((resolve) => {
-        directoryReader.readEntries((entries: Array<FileSystemFileEntry | FileSystemDirectoryEntry>) => {
-          resolve(entries);
-        });
+        directoryReader.readEntries(
+          (entries: Array<FileSystemFileEntry | FileSystemDirectoryEntry>) => {
+            resolve(entries);
+          }
+        );
       });
 
       return entries as Array<FileSystemFileEntry | FileSystemDirectoryEntry>;
@@ -271,14 +374,51 @@ export default class Volume extends Mixins(RefreshAppInfo) {
     }
   }
 
+  toggleModal(on: boolean): void {
+    if (!this.$refs.modal) return;
+    const modal = new Modal((this.$refs.modal as Vue).$el as HTMLElement, {});
+
+    on ? modal.show() : modal.hide();
+  }
+
+  async saveVolume(): Promise<void> {
+    const res = await authFetch(
+      `${this.$store.state.adminAPIURL}/admin/createOrUpdateVolume`,
+      {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+          id: this.volume._id,
+          name: this.volume.name,
+          files: JSON.stringify(this.fs.toJSON()),
+        }),
+      }
+    );
+
+    if (!res.ok) throw "Failed to save volume!";
+
+    this.volumeModified = false;
+  }
+
+  saveOpenFile(newFile: { name: string; contents: string}): void {
+    this.fs.writeFileSync(
+      this.currentDirectory + "/" + newFile.name,
+      newFile.contents
+    );
+    this.editing = false;
+    this.volumeModified = true;
+  }
+
   async mounted(): Promise<void> {
     await this.getVolumes();
 
     let id = this.$route.query.id;
-    
-    if (typeof id == 'string') {
+
+    if (typeof id == "string") {
       const vol = await this.getFullVolume(id);
-      
+
       if (vol) {
         this.volume = vol;
         this.fs = FSVolume.fromJSON(this.volume.files);
@@ -289,3 +429,9 @@ export default class Volume extends Mixins(RefreshAppInfo) {
   }
 }
 </script>
+
+<style>
+tr:not(.bg-primary) {
+  border-radius: 8px;
+}
+</style>
